requirements.txt
albumentationsx
pillow
matplotlib
scikit-learn
